Moves Common Item Holder Metadata Component to Abstract Class
Extended by inventory and trade variants.
